requirements.txt
fastapi==0.133.1
httpx==0.28.1
numpy==2.4.2
opencv_python==4.13.0.92
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.3
Requests==2.32.5
segment_anything==1.0
SQLAlchemy==2.0.47
# TensorRT Engine Building (Requires NVIDIA GPU)
# NOTE: Windows users must install via NVIDIA PyPI:
# pip install tensorrt --extra-index-url https://pypi.nvidia.com
tensorrt==10.15.1.29
torch==2.10.0
tqdm==4.67.3
ultralytics==8.4.18
uvicorn==0.34.2
python-multipart==0.0.22
psutil==5.9.8
onnxslim==0.1.89
onnxruntime-gpu==1.24.4
